+++ src/main/java/eu/clarin/linkchecker/bolt/MetricsFetcherBolt.java
@@ -508,7 +508,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                   // pass the info about denied by robots
                   metadata.setValue("fetch.category", Category.Blocked_By_Robots_txt.name());
                   collector.emit(com.digitalpebble.stormcrawler.Constants.StatusStreamName, fit.t,
-                        new Values(fit.url, metadata, Status.DISCOVERED));
+                        new Values(fit.url, metadata));
                   // no need to wait next time as we won't request from
                   // that site
                   asap = true;
@@ -545,7 +545,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                         metadata.setValue("fetch.category", Category.Blocked_By_Robots_txt.name());
 
                         collector.emit(com.digitalpebble.stormcrawler.Constants.StatusStreamName, fit.t,
-                              new Values(fit.url, metadata, Status.DISCOVERED));
+                              new Values(fit.url, metadata));
                         // no need to wait next time as we won't request
                         // from that site
                         asap = true;
@@ -580,7 +580,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                      metadata.setValue("fetch.category", Category.Restricted_Access.name());
                      
                      collector.emit(com.digitalpebble.stormcrawler.Constants.StatusStreamName, fit.t,
-                           new Values(fit.url, metadata, Status.DISCOVERED));
+                           new Values(fit.url, metadata));
                      
                      continue;
                   }
@@ -617,7 +617,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                            response.getMetadata().getFirstValue(HttpHeaders.LOCATION));
                      metadata.setValue("fetch.redirectCount", Integer.toString(redirectCount));
 
-                     collector.emit(eu.clarin.linkchecker.config.Constants.RedirectStreamName, fit.t, new Values(redirectUrl, metadata));
+                     collector.emit(eu.clarin.linkchecker.config.Constants.REDIRECT_STREAM_NAME, fit.t, new Values(redirectUrl, metadata));
                      continue;
                   }
                }
@@ -627,7 +627,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                // with the information that we wont to replay it with a GET request
                else if ("true".equals(metadata.getFirstValue("http.method.head")) && !Configuration.okStatusCodes.contains(response.getStatusCode())) {                  
                   metadata.setValue("http.method.head", "false");
-                  collector.emit(eu.clarin.linkchecker.config.Constants.RedirectStreamName, fit.t, new Values(fit.url, metadata));
+                  collector.emit(eu.clarin.linkchecker.config.Constants.REDIRECT_STREAM_NAME, fit.t, new Values(fit.url, metadata));
                   continue;
                }               
                else if (Configuration.okStatusCodes.contains(response.getStatusCode())) {
@@ -673,14 +673,14 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                mergedMD.setValue("fetch.redirectCount", Integer.toString(redirectCount));
 
                collector.emit(com.digitalpebble.stormcrawler.Constants.StatusStreamName, fit.t,
-                     new Values(fit.url, mergedMD, Status.DISCOVERED));
+                     new Values(fit.url, mergedMD));
 
             }
             catch (Exception exece) {
                // recheck with GET request if the failed check was a GET request
                if("true".equals(metadata.getFirstValue("http.method.head"))){
                   metadata.setValue("http.method.head", "false");
-                  collector.emit(eu.clarin.linkchecker.config.Constants.RedirectStreamName, fit.t, new Values(fit.url, metadata));
+                  collector.emit(eu.clarin.linkchecker.config.Constants.REDIRECT_STREAM_NAME, fit.t, new Values(fit.url, metadata));
                   continue;                  
                }
                
@@ -711,7 +711,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
                metadata.setValue("fetch.message", message);
 
                // send to status stream
-               collector.emit(Constants.StatusStreamName, fit.t, new Values(fit.url, metadata, Status.DISCOVERED));
+               collector.emit(Constants.StatusStreamName, fit.t, new Values(fit.url, metadata));
 
             }
             finally {
@@ -787,11 +787,11 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
    @Override
    public void declareOutputFields(OutputFieldsDeclarer declarer) {
       declarer.declareStream(
-            com.digitalpebble.stormcrawler.Constants.StatusStreamName, 
-            new Fields("url", "metadata", "status")
+            com.digitalpebble.stormcrawler.Constants.StatusStreamName,
+            new Fields("url", "metadata")
          );
       declarer.declareStream(
-            eu.clarin.linkchecker.config.Constants.RedirectStreamName, 
+            eu.clarin.linkchecker.config.Constants.REDIRECT_STREAM_NAME,
             new Fields("url", "metadata")
          );
    }
@@ -860,7 +860,7 @@ public class MetricsFetcherBolt extends StatusEmitterBolt {
          metadata.setValue("fetch.checkingDate", LocalDateTime.now().toString());
          
          collector.emit(com.digitalpebble.stormcrawler.Constants.StatusStreamName, input,
-               new Values(urlString, metadata, Status.ERROR));
+               new Values(urlString, metadata));
          collector.ack(input);
          return;
       }