+++ CHANGES.md
@@ -1,3 +1,8 @@
+# version 3.3.0
+- upgrading dependencies to Storm 2.6.1 and Storm Crawler 2.11  
+  => requirement to change user agent string (issues #78, #79, #80)
+- removing acknowledgement from MetricsFetcherBolt (issue #81)
+
 # version 3.2.0
 - adding functionality for host specific http.timeout
 - upgrading dependencies to Storm 2.5.0 and Storm Crawler 2.10